company 모듈을 생성하고 페이지네이션을 추가하였습니다.

+++ src/entities/common/common.entity.ts
@@ -1,7 +1,7 @@
 import { BaseEntity, CreateDateColumn, DeleteDateColumn, PrimaryGeneratedColumn, UpdateDateColumn } from 'typeorm';
 
 export class CommonEntity extends BaseEntity {
-  @PrimaryGeneratedColumn({ type: 'bigint' })
+  @PrimaryGeneratedColumn({ type: 'int' })
   id!: number;
 
   @CreateDateColumn()
